Module aws_sdk_kinesis::error

Modules

expired_iterator_exception

See ExpiredIteratorException

expired_next_token_exception

See ExpiredNextTokenException

invalid_argument_exception

See InvalidArgumentException

kms_access_denied_exception

See KmsAccessDeniedException

kms_disabled_exception

See KmsDisabledException

kms_invalid_state_exception

See KmsInvalidStateException

kms_not_found_exception

See KmsNotFoundException

kms_opt_in_required

See KmsOptInRequired

kms_throttling_exception

See KmsThrottlingException

limit_exceeded_exception

See LimitExceededException

provisioned_throughput_exceeded_exception

See ProvisionedThroughputExceededException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

Structs

AddTagsToStreamError

CreateStreamError

DecreaseStreamRetentionPeriodError

DeleteStreamError

DeregisterStreamConsumerError

DescribeLimitsError

DescribeStreamConsumerError

DescribeStreamError

DescribeStreamSummaryError

DisableEnhancedMonitoringError

EnableEnhancedMonitoringError

ExpiredIteratorException

The provided iterator exceeds the maximum age allowed.

ExpiredNextTokenException

The pagination token passed to the operation is expired.

GetRecordsError

GetShardIteratorError

IncreaseStreamRetentionPeriodError

InvalidArgumentException

A specified parameter exceeds its restrictions, is not supported, or can't be used.

KmsAccessDeniedException

The ciphertext references a key that doesn't exist or that you don't have access

KmsDisabledException

The request was rejected because the specified customer master key (CMK) isn't

KmsInvalidStateException

The request was rejected because the state of the specified resource isn't valid

KmsNotFoundException

The request was rejected because the specified entity or resource can't be

KmsOptInRequired

The AWS access key ID needs a subscription for the service.

KmsThrottlingException

The request was denied due to request throttling. For more information about

LimitExceededException

The requested resource exceeds the maximum number allowed, or the number of

ListShardsError

ListStreamConsumersError

ListStreamsError

ListTagsForStreamError

MergeShardsError

ProvisionedThroughputExceededException

The request rate for the stream is too high, or the requested data is too large for

PutRecordError

PutRecordsError

RegisterStreamConsumerError

RemoveTagsFromStreamError

ResourceInUseException

The resource is not available for this operation. For successful operation, the

ResourceNotFoundException

The requested resource could not be found. The stream might not be specified

SplitShardError

StartStreamEncryptionError

StopStreamEncryptionError

UpdateShardCountError

Enums

AddTagsToStreamErrorKind

CreateStreamErrorKind

DecreaseStreamRetentionPeriodErrorKind

DeleteStreamErrorKind

DeregisterStreamConsumerErrorKind

DescribeLimitsErrorKind

DescribeStreamConsumerErrorKind

DescribeStreamErrorKind

DescribeStreamSummaryErrorKind

DisableEnhancedMonitoringErrorKind

EnableEnhancedMonitoringErrorKind

GetRecordsErrorKind

GetShardIteratorErrorKind

IncreaseStreamRetentionPeriodErrorKind

ListShardsErrorKind

ListStreamConsumersErrorKind

ListStreamsErrorKind

ListTagsForStreamErrorKind

MergeShardsErrorKind

PutRecordErrorKind

PutRecordsErrorKind

RegisterStreamConsumerErrorKind

RemoveTagsFromStreamErrorKind

SplitShardErrorKind

StartStreamEncryptionErrorKind

StopStreamEncryptionErrorKind

UpdateShardCountErrorKind