Module aws_sdk_kinesis::model

Modules

child_shard

See ChildShard

consumer

See Consumer

consumer_description

See ConsumerDescription

enhanced_metrics

See EnhancedMetrics

hash_key_range

See HashKeyRange

put_records_request_entry

See PutRecordsRequestEntry

put_records_result_entry

See PutRecordsResultEntry

record

See Record

sequence_number_range

See SequenceNumberRange

shard

See Shard

shard_filter

See ShardFilter

stream_description

See StreamDescription

stream_description_summary

See StreamDescriptionSummary

tag

See Tag

Structs

ChildShard

Consumer

An object that represents the details of the consumer you registered. This type of

ConsumerDescription

An object that represents the details of a registered consumer. This type of object is

EnhancedMetrics

Represents enhanced metrics types.

HashKeyRange

The range of possible hash key values for the shard, which is a set of ordered

PutRecordsRequestEntry

Represents the output for PutRecords.

PutRecordsResultEntry

Represents the result of an individual record from a PutRecords

Record

The unit of data of the Kinesis data stream, which is composed of a sequence

SequenceNumberRange

The range of possible sequence numbers for the shard.

Shard

A uniquely identified group of data records in a Kinesis data stream.

ShardFilter

StreamDescription

Represents the output for DescribeStream.

StreamDescriptionSummary

Represents the output for DescribeStreamSummary

Tag

Metadata assigned to the stream, consisting of a key-value pair.

Enums

ConsumerStatus

EncryptionType

MetricsName

ScalingType

ShardFilterType

ShardIteratorType

StreamStatus