Module aws_sdk_kinesis::operation

Structs

AddTagsToStream

Adds or updates tags for the specified Kinesis data stream. Each time you invoke

CreateStream

Creates a Kinesis data stream. A stream captures and transports data records that

DecreaseStreamRetentionPeriod

Decreases the Kinesis data stream's retention period, which is the length of time

DeleteStream

Deletes a Kinesis data stream and all its shards and data. You must shut down any

DeregisterStreamConsumer

To deregister a consumer, provide its ARN. Alternatively, you can provide the ARN of

DescribeLimits

Describes the shard limits and usage for the account.

DescribeStream

Describes the specified Kinesis data stream.

DescribeStreamConsumer

To get the description of a registered consumer, provide the ARN of the consumer.

DescribeStreamSummary

Provides a summarized description of the specified Kinesis data stream without the

DisableEnhancedMonitoring

Disables enhanced monitoring.

EnableEnhancedMonitoring

Enables enhanced Kinesis data stream monitoring for shard-level metrics.

GetRecords

Gets data records from a Kinesis data stream's shard.

GetShardIterator

Gets an Amazon Kinesis shard iterator. A shard iterator expires 5 minutes after it

IncreaseStreamRetentionPeriod

Increases the Kinesis data stream's retention period, which is the length of time

ListShards

Lists the shards in a stream and provides information about each shard. This

ListStreamConsumers

Lists the consumers registered to receive data from a stream using enhanced fan-out,

ListStreams

Lists your Kinesis data streams.

ListTagsForStream

Lists the tags for the specified Kinesis data stream. This operation has a limit of

MergeShards

Merges two adjacent shards in a Kinesis data stream and combines them into a single

PutRecord

Writes a single data record into an Amazon Kinesis data stream. Call

PutRecords

Writes multiple data records into a Kinesis data stream in a single call (also

RegisterStreamConsumer

Registers a consumer with a Kinesis data stream. When you use this operation, the

RemoveTagsFromStream

Removes tags from the specified Kinesis data stream. Removed tags are deleted and

SplitShard

Splits a shard into two new shards in the Kinesis data stream, to increase the

StartStreamEncryption

Enables or updates server-side encryption using an AWS KMS key for a specified

StopStreamEncryption

Disables server-side encryption for a specified stream.

UpdateShardCount

Updates the shard count of the specified stream to the specified number of