Module aws_sdk_kinesis::error

Modules

expired_iterator_error
See ExpiredIteratorError

expired_next_token_error
See ExpiredNextTokenError

invalid_argument_error
See InvalidArgumentError

kms_access_denied_error
See KMSAccessDeniedError

kms_disabled_error
See KMSDisabledError

kms_invalid_state_error
See KMSInvalidStateError

kms_not_found_error
See KMSNotFoundError

kms_opt_in_required
See KMSOptInRequired

kms_throttling_error
See KMSThrottlingError

limit_exceeded_error
See LimitExceededError

provisioned_throughput_exceeded_error
See ProvisionedThroughputExceededError

resource_in_use_error
See ResourceInUseError

resource_not_found_error
See ResourceNotFoundError

Structs

AddTagsToStreamError

CreateStreamError

DecreaseStreamRetentionPeriodError

DeleteStreamError

DeregisterStreamConsumerError

DescribeLimitsError

DescribeStreamConsumerError

DescribeStreamError

DescribeStreamSummaryError

DisableEnhancedMonitoringError

EnableEnhancedMonitoringError

ExpiredIteratorError
The provided iterator exceeds the maximum age allowed.

ExpiredNextTokenError
The pagination token passed to the operation is expired.

GetRecordsError

GetShardIteratorError

IncreaseStreamRetentionPeriodError

InvalidArgumentError
A specified parameter exceeds its restrictions, is not supported, or can't be used.

KMSAccessDeniedError
The ciphertext references a key that doesn't exist or that you don't have access

KMSDisabledError
The request was rejected because the specified customer master key (CMK) isn't

KMSInvalidStateError
The request was rejected because the state of the specified resource isn't valid

KMSNotFoundError
The request was rejected because the specified entity or resource can't be

KMSOptInRequired
The AWS access key ID needs a subscription for the service.

KMSThrottlingError
The request was denied due to request throttling. For more information about

LimitExceededError
The requested resource exceeds the maximum number allowed, or the number of

ListShardsError

ListStreamConsumersError

ListStreamsError

ListTagsForStreamError

MergeShardsError

ProvisionedThroughputExceededError
The request rate for the stream is too high, or the requested data is too large for

PutRecordError

PutRecordsError

RegisterStreamConsumerError

RemoveTagsFromStreamError

ResourceInUseError
The resource is not available for this operation. For successful operation, the

ResourceNotFoundError
The requested resource could not be found. The stream might not be specified

SplitShardError

StartStreamEncryptionError

StopStreamEncryptionError

UpdateShardCountError

Enums

AddTagsToStreamErrorKind

CreateStreamErrorKind

DecreaseStreamRetentionPeriodErrorKind

DeleteStreamErrorKind

DeregisterStreamConsumerErrorKind

DescribeLimitsErrorKind

DescribeStreamConsumerErrorKind

DescribeStreamErrorKind

DescribeStreamSummaryErrorKind

DisableEnhancedMonitoringErrorKind

EnableEnhancedMonitoringErrorKind

GetRecordsErrorKind

GetShardIteratorErrorKind

IncreaseStreamRetentionPeriodErrorKind

ListShardsErrorKind

ListStreamConsumersErrorKind

ListStreamsErrorKind

ListTagsForStreamErrorKind

MergeShardsErrorKind

PutRecordErrorKind

PutRecordsErrorKind

RegisterStreamConsumerErrorKind

RemoveTagsFromStreamErrorKind

SplitShardErrorKind

StartStreamEncryptionErrorKind

StopStreamEncryptionErrorKind

UpdateShardCountErrorKind